--- a/builder/builder.go
+++ b/builder/builder.go
@@ -12,13 +12,15 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package redact
+package builder
 
 import (
-	"bytes"
 	"fmt"
 	"io"
-	"unicode/utf8"
+
+	i "github.com/cockroachdb/redact/interfaces"
+	ib "github.com/cockroachdb/redact/internal/buffer"
+	ifmt "github.com/cockroachdb/redact/internal/rfmt"
 )
 
 // StringBuilder accumulates strings with optional redaction markers.
@@ -27,26 +29,13 @@ import (
 // To distinguish safe and unsafe bits, it also implements the SafeWriter
 // interface.
 type StringBuilder struct {
-	// we use bytes.Buffer internally to simplify the implementation of
-	// the SafeWriter interface.
-	buf bytes.Buffer
+	ib.Buffer
 }
 
-// String returns the accumulated string, with redaction markers stripped out.
-// To obtain the redactable string, call RedactableString().
-func (b StringBuilder) String() string { return b.RedactableString().StripMarkers() }
-
-var _ fmt.Stringer = StringBuilder{}
 var _ fmt.Stringer = (*StringBuilder)(nil)
 
-// RedactableString returns the accumulated string, including redaction markers.
-func (b StringBuilder) RedactableString() RedactableString { return RedactableString(b.buf.String()) }
-
-// RedactableBytes returns the accumulated bytes, including redaction markers.
-func (b StringBuilder) RedactableBytes() RedactableBytes { return RedactableBytes(b.buf.Bytes()) }
-
 // SafeFormat implements SafeFormatter.
-func (b StringBuilder) SafeFormat(p SafePrinter, _ rune) {
+func (b StringBuilder) SafeFormat(p i.SafePrinter, _ rune) {
 	// We only support the %v / %s natural print here.
 	// Go supports other formatting verbs for strings: %x/%X/%q.
 	//
@@ -64,20 +53,8 @@ func (b StringBuilder) SafeFormat(p SafePrinter, _ rune) {
 	p.Print(b.RedactableString())
 }
 
-var _ SafeFormatter = StringBuilder{}
-var _ SafeFormatter = (*StringBuilder)(nil)
-
-// Len returns the number of accumulated bytes, including redaction
-// markers; b.Len() == len(b.RedactableString()).
-func (b *StringBuilder) Len() int { return b.buf.Len() }
-
-// Cap returns the capacity of the builder's underlying byte slice. It is the
-// total space allocated for the string being built and includes any bytes
-// already written.
-func (b *StringBuilder) Cap() int { return b.buf.Cap() }
-
-// Reset resets the Builder to be empty.
-func (b *StringBuilder) Reset() { b.buf.Reset() }
+var _ i.SafeFormatter = StringBuilder{}
+var _ i.SafeFormatter = (*StringBuilder)(nil)
 
 // StringBuilder implements io.Writer.
 // Direct Write() calls are considered unsafe.
@@ -85,65 +62,57 @@ var _ io.Writer = (*StringBuilder)(nil)
 
 // Write implements the io.Writer interface.
 func (b *StringBuilder) Write(s []byte) (int, error) {
-	b.UnsafeBytes(s)
-	return len(s), nil
+	b.SetMode(ib.UnsafeEscaped)
+	return b.Buffer.Write(s)
 }
 
 // StringBuilder implements SafeWriter.
-var _ SafeWriter = (*StringBuilder)(nil)
+var _ i.SafeWriter = (*StringBuilder)(nil)
 
 // Print is part of the SafeWriter interface.
 func (b *StringBuilder) Print(args ...interface{}) {
-	_, _ = Fprint(&b.buf, args...)
+	b.SetMode(ib.PreRedactable)
+	_, _ = ifmt.Fprint(&b.Buffer, args...)
 }
 
 // Printf is part of the SafeWriter interface.
 func (b *StringBuilder) Printf(format string, args ...interface{}) {
-	_, _ = Fprintf(&b.buf, format, args...)
+	b.SetMode(ib.PreRedactable)
+	_, _ = ifmt.Fprintf(&b.Buffer, format, args...)
 }
 
 // SafeString is part of the SafeWriter interface.
-func (b *StringBuilder) SafeString(s SafeString) {
-	w := escapeWriter{w: &b.buf, enclose: false}
-	_, _ = w.Write([]byte(s))
+func (b *StringBuilder) SafeString(s i.SafeString) {
+	b.SetMode(ib.SafeEscaped)
+	_, _ = b.Buffer.WriteString(string(s))
 }
 
 // SafeRune is part of the SafeWriter interface.
-func (b *StringBuilder) SafeRune(s SafeRune) {
-	if s == startRedactable || s == endRedactable {
-		s = escapeMark
-	}
-	_, _ = b.buf.WriteRune(rune(s))
+func (b *StringBuilder) SafeRune(s i.SafeRune) {
+	b.SetMode(ib.SafeEscaped)
+	_ = b.Buffer.WriteRune(rune(s))
 }
 
 // UnsafeString is part of the SafeWriter interface.
 func (b *StringBuilder) UnsafeString(s string) {
-	w := escapeWriter{w: &b.buf, enclose: true, strip: true}
-	_, _ = w.Write([]byte(s))
+	b.SetMode(ib.UnsafeEscaped)
+	_, _ = b.Buffer.WriteString(s)
 }
 
 // UnsafeRune is part of the SafeWriter interface.
 func (b *StringBuilder) UnsafeRune(s rune) {
-	_, _ = b.buf.WriteRune(startRedactable)
-	b.SafeRune(SafeRune(s))
-	_, _ = b.buf.WriteRune(endRedactable)
+	b.SetMode(ib.UnsafeEscaped)
+	_ = b.Buffer.WriteRune(s)
 }
 
 // UnsafeByte is part of the SafeWriter interface.
 func (b *StringBuilder) UnsafeByte(s byte) {
-	_, _ = b.buf.WriteRune(startRedactable)
-	if s >= utf8.RuneSelf ||
-		s == startRedactableS[0] || s == endRedactableS[0] {
-		// Unsafe byte. Escape it.
-		_, _ = b.buf.Write(escapeBytes)
-	} else {
-		_ = b.buf.WriteByte(s)
-	}
-	_, _ = b.buf.WriteRune(endRedactable)
+	b.SetMode(ib.UnsafeEscaped)
+	_ = b.Buffer.WriteByte(s)
 }
 
 // UnsafeBytes is part of the SafeWriter interface.
 func (b *StringBuilder) UnsafeBytes(s []byte) {
-	w := escapeWriter{w: &b.buf, enclose: true, strip: true}
-	_, _ = w.Write(s)
+	b.SetMode(ib.UnsafeEscaped)
+	_, _ = b.Buffer.Write(s)
 }