… bench_test.go
… builder/builder_test.go
… internal/buffer/buffer_test.go
… internal/escape/escape.go
… internal/escape/escape_test.go
… internal/fmtforward/make_format_test.go
… internal/hooks.go
… internal/markers/constants.go
… internal/markers/markers.go
… internal/print.go.diff
… internal/redact/doc.go
… internal/redact/wrappers.go
… internal/rfmt/README.md
… internal/rfmt/fmtsort/sort.go
… internal/rfmt/format.go
… internal/rfmt/format.go.diff
… internal/rfmt/helpers.go
… internal/rfmt/print.go
… internal/rfmt/print.go.diff
… internal/rfmt/printer_adapter.go
… internal/rfmt/refresh.sh
… internal/rfmt/registry.go
… internal/rfmt/registry_test.go
… markers_internal_constants.go
… markers_internal_escape.go
… markers_internal_print.go
… markers_internal_printer.go
… markers_print.go
… markers_test.go
… util.go