Fixes #5 
Fixes #15 

**tldr: this change preserves the existing APIs but makes the library
faster and also able to recursively format structs and arrays.**

Also a new API: `redact.ManualBuffer`, which gives ownership of
swapping between safe and unsafe output to the caller. This is
envisioned to be useful in CockroachDB to help implement `tree.FmtCtx`.

Details:

We are really looking at a major redesign.

The previous design was attempting to preserve Go's own `fmt/print.go`
as much as possible, and was "injecting" special formatting behavior
by wrapping every printed value in a custom object whose type had a
`Format()` method. The idea was to give “ownership” of adding
redaction markers to each argument in turn. That worked for simple
values but there was no good way to make it able to print structs and
arrays recursively.

This new implementation customizes the `print.go` code more heavily.
In particular, the check for safe/unsafe types is now done at
every level of the recursive print. It also makes the code use a
`redact.ManualBuffer` as target during the print evaluation,
which saves string copies (hence a general performance boost).

The new supported cases can be understood by looking at the diffs
on the files `markers_test.go` and `internal/rfmt/registry_test.go`.


Note about the recursive display of struct values: the print logic
uses reflection to print out structs. However, the Go semantics
prevent a package from looking at non-exported fields in structs from
other packages. This means that the print logic is unable to recognize
the various safety interfaces for non-exported struct fields:
`SafeValue`, `SafeFormatter`, `FormatError`, etc. These are thus
considered unsafe regardless of the interfaces implemented.

The safe type registry and the special types `RedactableString` and
`RedactableBytes` are still recognized for non-exported structs
though.

Peformance improvement using the included `bench_test.go`:

```
name       old time/op    new time/op    delta
Redact-32    2.71µs ±22%    1.49µs ± 0%  -45.25%  (p=0.000 n=10+9)

name       old alloc/op   new alloc/op   delta
Redact-32      360B ± 0%      248B ± 0%  -31.11%  (p=0.000 n=10+10)

name       old allocs/op  new allocs/op  delta
Redact-32      12.0 ± 0%       5.0 ± 0%  -58.33%  (p=0.000 n=10+10)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/redact/19)
<!-- Reviewable:end -->