@knz 2021-06-19
cc @tbg  - long awaited! cost ~3 days of work.

I'm not expecting anyone to be able to review this unless they have already intimate knowledge of Go's `fmt` internals.

Instead:

- I was careful to ensure that the external API of the package has not changed, so we can compare before/after on downstream packages without code changes
- I ensured that the `cockroachdb/errors` library's own test suite passes with this new code without changes. As the `errors` library has a pretty extensive and detailed formatting test suite, that should give us confidence that there are no regressions.
- I'll wait until we re-enable Travis CI and that gives green before merging the branch.
- I'll ensure this is published with a new "major" version (v1.1, we're at v1.0 currently) and clarify there may be unforeseen breakage in the release notes.